Updated Version and some art
Art is in the code as byte array, so not needed, but here to look at.

Difficulty working and added an end screen. No animation and testing shortcut still present.
Not finished but all added aspects working. Requires animation and removing of testing shortcut. Also think using real and string for difficulty is poor (easy to get wrong, messy and uses fair bit of memory), will move to an int and reference to mke look nicer in code method.